requirements.txt
# Python 依赖
DrissionPage>=4.0.0
PyYAML>=6.0
requests>=2.28.0
